mxsboot: added mxsboot as native code for update
This is a host tool provided to generate a bootable image
for NAND/MMC.
We need to use some of the functions included here to patch
our implementation of kobs-ng (cmd_bootstream.c) which leaned
on NAND driver code which no longer exists. These are the
functions to set the bad block marker byte and offset in the
FCB block.

https://jira.digi.com/browse/DUB-123

Signed-off-by: Hector Palacios <hector.palacios@digi.com>