bootstream: new calculation of bad block marker byte/offset
The old function leaned on the old NAND driver to retrieve the
information about the bad block marker byte and bit offset.
This no longer exists, but the function is implemented in mxsboot.c
which we brought in and now compile as native code (with the
exception of functions done for the host app).

https://jira.digi.com/browse/DUB-123

Signed-off-by: Hector Palacios <hector.palacios@digi.com>