digi/env: support CONFIG_ENV_SIZE changes
If U-Boot has a built-in CONFIG_ENV_SIZE that doesn't match
the size of the environment saved in NVRAM (typically when
using a new U-Boot on a module with old-U-Boot generated NVRAM), we
want to modify the U-Boot env partition in NVRAM (always the
last entry) with the new built-in size.
This way, if the ENV data on the old NVRAM was ok, it is preserved
but the partition updated with the new size (only if a 'saveenv'
is run).

https://jira.digi.com/browse/DUB-145

Signed-off-by: Hector Palacios <hector.palacios@digi.com>