ccardimx28: init only the MAC address of FEC1
For this platform we will only support the FEC0.
FEC1 cannot work without it (as it communicates through FEC0
MDIO bus). We don't want to suppport this interface in U-Boot
except for the MAC address initialization.
A config setting is provided so that FEC1 is only initialized
with the MAC address.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>