board: digi: added support for Device Tree partitions
The Device Tree Blob is a file with extension *.dtb that
contains a Flattened Device Tree needed to boot modern Linux
kernels. This file must be readable by U-Boot from flash and
the easiest approach (currently) is to have a dedicated raw
partition for it in the Flash.

This is considered a 'system' partition (just like the kernel
and rootfs partitions) as far as dual boot mechanism concerns.

https://jira.digi.com/browse/DUB-150

Signed-off-by: Hector Palacios <hector.palacios@digi.com>