bsp: digi: platform name may be composed or not
Previous commit 692e3e8bb72072130444bebd7295acc7a893365c
introduced a mechanism to dynamically generate the platform name.
This assumed that all Digi platforms are composed as a combination of
prefix + [i|w] + suffix + platform
which is not true for some platforms like CPX2 or WR21.

This commit adds a new define called MODULE_COMPOSED_NAME so
that composed platform names are generated like that, but other
platforms with fixed names, are built out of the CONFIG_MODULE_NAME
constant.

It also fixes a bug when adding the 'w' to wireless platforms
where only the imx53 platforms were considered.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Javier Viguera <javier.viguera@digi.com>