dualboot: reset/retry forever on CRC/header failure
When booting a Linux image U-Boot checks the header and CRC
of the image and it is able to determine if the image is bootable.
When it isn't, and dual boot is enabled, it will immediately mark
the system as invalid and toggles to the alternate partition or
calls the panic hook. This is independent on the config switch
CONFIG_DUAL_BOOT_RETRY_FOREVER, because in this case, it doesn't
make much sense to keep retrying on a firmware that U-Boot itself
has verified it cannot boot.

This case however could be hit by a transient NAND read problem
that a reset might solve.
Apart from that, some people believe that the if the RETRY_FOREVER
option is selected, U-Boot should retry forever, even on verified
bad firmware.

This patch solves the transient NAND read case and modifies the
current behaviour to this when U-Boot detects a corrupt firmware:

1. It will increment the boot attempts counter
2. If the max retries limit is not reached, it will reset the board
   (to catch potential transient NAND read errors)
3. If the limit is reached, it will toggle to the alternate partition
   or else:
   - It will retry forever on the invalid firmware if
     CONFIG_DUAL_BOOT_RETRY_FOREVER is set, or
   - it will call the panic hook, otherwise

https://jira.digi.com/browse/DUB-170

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Robert Hodaszi <robert.hodaszi@digi.com>