bsp: digi: dynamic calculation of 'loadaddr' and similar
A previous commit 3c6520a0423c7eb315f2fc572e98fdbb6db5308f
did a dynamic setting of constant values for 'loadaddr' and
'wceloadaddr'. This commit was ugly and incomplete due to the
difficulty to do calculations in preprocessing code.

This new patch removes all the constant values and logic from
the platform headers and moves the logic to the platform code.
Only CONFIG_LOADADDR remains (used as default load address for
some commands), and is set to a low memory address (8MB).

The new function get_platform_loadaddr() is passed the name of the
address to calculate (currently only 'loadaddr' and 'wceloadaddr')
and returns the most appropriate value depending on the available
RAM of the platform. The idea behind is to allow TFTP transfers and
WinCE boots of the largest possible firmware image, without needing
to rely on On-The-Fly update mechanism.

https://jira.digi.com/browse/DUB-172

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Robert Hodaszi <robert.hodaszi@digi.com>