bsp: digi: activate OTF if update partition bigger than available RAM
In previous commit standard TFTP update has returned to be the default
behavior and On-the-fly update must be manually enabled.
If the variable otf-update is not defined, OTF was auto activated when
updating rootfs or android partitions which are usually large.
This approach was not very exact and many times OTF was used on platforms
which should have enough RAM memory to do large file updates.
To fix this, since OTF has the drawback of first erasing the flash
before transferring the firmware, this patch auto activates OTF update
when the partition to update is bigger than the available RAM memory
for doing TFTP transfers (notice we talk about partition size and not
file size, because with TFTP the filesize is unkown until completely
transferred).

This approach now lets U-Boot use the standard TFTP update method (first
transfer, then erase, then write), in most cases and relegates OTF
mechanism to extraordinary cases (of large files and small RAM memory
variants).

https://jira.digi.com/browse/DUB-172

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Robert Hodaszi <robert.hodaszi@digi.com>