bsp: digi: for update/verify commands, load firmware at 'loadaddr'
The distinction between the different load addresses only makes sense
on the dboot command. For example, WinCE images need to be transferred
to a location in RAM that later allows to uncompress the image to another
location. For updating firmware or verifying it, we can use the standard
loadaddres, which is calculated to allow the largest possible transfer
depending on available RAM memory.

https://jira.digi.com/browse/DUB-172

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Robert Hodaszi <robert.hodaszi@digi.com>