bsp: digi: do not apend padding when image size aligned to page
The 'update' command was always appending a padding to images to
align the data to write to the flash to a page size (for example 2K).
This was done even when the data to write was already aligned.
If the partition was larger than the data size it was not a problem
but if the data size (firmware file) is exactly as big as the partition
the 'update' command failed due to partition smaller than data size.

https://jira.digi.com/browse/DUB-176

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Javier Viguera <javier.viguera@digi.com>