board: ccardimx28: default to 256MB if invalid variant on OTP
When the HWID variant has not been coded in the OTP bits, a zero
is read and return as RAM size, which is an invalid value and
U-Boot refuses to start. This patch defaults to 256MB RAM size
(JSK module -variant 0x02 default) if the information on the OTP
is invalid.

https://jira.digi.com/browse/DUB-181

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>