board: digi: ccardimx28: register HWID into the Device Tree
The proper way to pass hardware information to the kernel
is through the Device Tree, so we'll encode the HWID as properties
of the main node. Properties will be prefixed with string "digi,hwid,"
followed by the name of each property:
     digi.hwid,tf
     digi,hwid,variant,
     digi,hwid,hv
     digi,hwid,cert
     digi,hwid,year
     digi,hwid,month
     digi,hwid,sn

The boot loader will read the HWID from the OTP bits (or from NVRAM)
and embed these properties in the DT before booting the kernel.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Javier Viguera <javier.viguera@digi.com>

https://jira.digi.com/browse/DUB-184