board: ccardimx28: handle different polarities for FET enabling
On modules newer than V1, a FET is blocking the 3V3 supply.
To have this voltage enabled to the module LCD_RS
must be driven low on modules V2..V5 and driven high on modules V6.
On V1 modules this line is not used so the code is harmless.
Modules with blank OTP bits are by default considered as V6.

The code has been moved from board_early_init_f() to board_late_init()
because it now depends on the hardware variant coded in the HWID
which is not read until a later point (when NAND has been initialized).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Robert Hodaszi <robert.hodaszi@digi.com>

https://jira.digi.com/browse/DUB-186