board: digi: reuse existing ft_board_setup to update FDT
Commit c6b90e7aee699d82b555f7c1ea3faec8727d5dc3 introduced a new
function to modify the FDT before booting with board custom code.
There already exist a common function for this which is called
ft_board_setup() and that is from 'bootm' command.
Update the code to reuse this function instead and modify the FDT
using API function do_fixup_by_path() instead of 'fdt' command.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Robert Hodaszi <robert.hodaszi@digi.com>