ccimx6js: move U-Boot environmet to eMMC BootPart2
Since U-Boot will be located in eMMC boot partition 1, let's
move the environment to boot partition 2 (which would otherwise
be unused). This way we offer a some extra protection to the
environment, making sure it cannot be accidentally overwritten
when U-Boot is updated.

The first copy will be located at offset 1MB (to allow for space for
a possible secondary bootloader on the boot partition 2), the redundant
copy will be located 128Kb after the first copy.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>