Revert "part_efi: zero-fill the protective MBR struct initially"
This reverts commit 1403cfd75c8dd627345931b999b35677f31dd198.
The calloc() calls allocs and memsets the area to zero.
The problem why we saw data in the MBR is when we boot U-Boot from
uSD card (environment is saved in 1MB offset) and we try to partition
the same uSD card we've booted from. This access conflicts with the
U-Boot environment and may hang U-Boot.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>