ccimx6: helper string to calculate $filesize in blocks
When writing to the eMMC, we need to specify the amount of data to
write in 512byte blocks.
When a file is transferred, U-Boot automatically sets $filesize to
the number of bytes of the file. We need to divide this by 512 and
add 1 to get the number of 512byte blocks that we need to write to
the eMMC.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>