ccimx6: append contents of 'bootargs_once' to bootargs
On the scripts that set the 'bootargs' variable before booting Android
or Linux, append the contents of variable 'bootargs_once'.
This variable name has been reserved as read-only string, so it cannot
be set as U-Boot variable. It is intended to be a hush local,
non-permanent variable instead whose contents get reset on the next
reboot cycle.

We will use this to set specific bootargs that can only be used on the
current run, after they have been set in U-Boot with 'variable=value'.
For example to format Android's data and cache partitions during the first
run (manufacturing).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>