cmd_update: Added Digi 'update' command
update command allows you to define the key name or partition name
to update, the media from where to get the firmware file, the device
index:partition, the file system, and the file name to use.

This first version only recognizes on key name and default filename:

	Key name		Filename variable
	--------		-----------------
	uboot		$uboot_file

Other than that a GPT partition name must be provided together with
the firmware file.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>