update: added OTF update mechanism for TFTP
The OTF (on-the-fly) update mechanism registers a board-code hook function
that is called by the transfer media (in this patch only TFTP is
supported) so that the bytes (transferred in packets) are buffered to
RAM, and later written to the media in chunks of a specific size
(now determined by constant CONFIG_OTF_CHUNK).

This allows to do firmware update without requiring too much free RAM
memory.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-278