ccimx6: if undefined, autogenerate 'mmcdev' basing on boot device
Users will expect that the system is located on the same media that the
target is booting from (if booting from eMMC, on eMMC; if booting from
uSD card, on uSD card).
To that end, the bootargs must generate a 'root=/dev/mmcblkXpY' where X
is the boot device.
This patch autogenerates 'mmcdev' if it doesn't exist, to point to the
boot device.
By default, do not define 'mmcdev'.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-283