ccimx6: simplify OTF update hook function
Remove some internal variables and rework the code to better fit for other
media like MMC.
Reset internal variables during the FLUSH operation.
Now, we first copy the buffer to RAM (unless 'dst' and 'loadaddr' are
the same, which would mean the buffer is already in RAM).
When the size of the buffer exceeds CONFIG_OTF_CHUNK, we write as many
blocks as we can to the eMMC. This may be larger than CONFIG_OTR_CHUNK
actually, so this constant should not be a tight limit of the RAM.

For TFTP, we don't need to provide the 'offset' where the chunk should go
as this is implicit by the 'len' parameter (the hook accumulates the
chunk length and knows at any moment the offset where the chunk should
be placed). For MMC, the offset parameter will be used in the other
direction: the hook will pass here the offset where the block driver must
read the next chunk to ($loadaddr) to reuse the memory.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>