update: added support for on-the-fly (OTF) update from MMC
We must call the otf hook on the MMC block read function. This function
reads a large number of blocks in one go (in ccimx6 65535 blocks, which
represent almost 32MiB).
The otfd.offset parameter is modified by the hook to tell the caller
(here the mmc driver) the RAM position where he should place the read
data to.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>