ccimx6: default bootcmd should run boot script from MMC
By default, we want our kits to be able to run Android and Yocto from
either eMMC or uSD card out of the box (without user intervention).
The problem is Android and Linux use different 'bootargs' and the root
partition in eMMC may be different from the root partition in uSD cards
as created from standard build systems. To overcome this problem, this
patch sets the $bootcmd by default to run the $loadscript which reads
$script from the $mmcdev:$mmcpart (eMMC or uSD, depending on the boot
media). If it exists, it runs it.

Android and Linux may create custom scripts to modify on-the-fly certain
variables (for example $mmcroot) before running the 'dboot' command, so
that they fit the expected 'bootargs' when booting from the uSD card.
If no script exists, the default variables are prepared to boot from the
eMMC ($mmcroot=$part3_uuid).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-285