update: get default filename from variable called after partition
If a filename is not provided, update command will get the value of a
variable that's named after the partition name and the suffix '_file'.
For example, for a partition called 'data' it will try to get the filename
from variable 'data_file', for 'system2', 'system2_file' and so on.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-287