ccimx6: hardcode 'mmcroot' to part3 UUID
In commit c07ec72248d96d34cf420683d4e5d5fb522920c7 we were
setting ${mmcroot} to ${part3_uuid}, but this is later not
expanded twice when setting the bootargs for Linux, so the
kernel received the string ${part3_uuid} which means nothing.

To fix it, set ${mmcroot} to the UUID string of partition 3.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-301