fec_mxc: do not set the MAC from OTP bits if so configured
Added config option CONFIG_NO_MAC_FROM_OTP to avoid having the FEC driver
set the MAC from the OTP bits.
In Digi platforms, we are not using the OTP bits for the MAC address.
This avoids the warning message about MAC being different in the
environment than in the SOM.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-305