ccimx6: move back setup_pmic_voltages() to board code
The ADPT and SBC boards may have differences in the power sources and
gpios of the PMIC, so let's move back the setup_pmic_voltages() function
to board code.
This patch also creates a header file to contain the definitions of the
common functions and board-specific functions that are used by the common
and board-specific code.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>