dboot/update: create macros to build digi commands help text
Both 'dboot' and 'update' commands have similar syntax and arguments.
We will implement a new command 'updatefile' that will also be similar.
By creating some constant strings, we can reuse them to build consistent
help for the commands and avoid duplicated strings and efforts when it
comes to change the text.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>