ccimx6: remove 'bpp=32' from video variables
On kernel v3.10 this is passed through the DT.
On Android kernel v3.0 this has been set as the default
if not passed through the cmdline.

Notice that this parameter can still be passed through the
command line to override the default setting or the setting
on the DT. The decision to remove it from the 'video'
variables responds to not influence customers to try other
'bpp' values, which we haven't tested.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>