ccimx6: disable PMIC backlight when enbling 5V source
PMIC GPIO11 is the LCD backlight which is low level enabled.
If left with default configuration (input) or when coming from
power-off, the backlight may be enabled and draw too much
power from the 5V source when this source is enabled, which
may cause a voltage drop on the 5V line and hang the I2C bus
where the touch controller is attached.

To prevent this, this patch configures GPIO11 as output and
sets it high, to make sure the backlight is disabled when the
5V source is enabled.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
Reviewed-by: Alex Gonzalez <alex.gonzalez@digi.com>

https://jira.digi.com/browse/ADK4A-153