env: process -f (forcibly) option of 'env default' command
The 'env default' command offers the -f option to forcibly reset the
environment to its defaults. This option is however not taken into account
with the following inconsistent behavior for protected variables:

When resetting a single variable:
  env default ethaddr      => Does not change 'ethaddr' (OK)
  env default -f ethaddr   => Does not change 'ethaddr' (Problem #1)

When resetting the whole environment:
  env default -a           => Changes/deletes 'ethaddr' (Problem #2)
  env default -f -a        => Changes/deletes 'ethaddr' (OK)

For Problem #1, this patch propagates the -f flag to set_default_vars().
For Problem #2, this patch calls set_default_env() function with a star (*)
as the first character of the passed string if called with -f.
This is later processed to not pass the flag H_NOCLEAR (passed now by
default) to h_import() with the following effects:
 1) normal existing variables are reset to their default value.
 2) protected existing variables are modified only if the -f option is
    used.
 3) existing variables without an assigned default value are preserved.

Effect #3 may be undesired but it is less critical than the current
behavior where protected variables can be modified by resetting to the
default environment.
In the future we may want to have an additional option -c (clear) that
can clear normal variables that did not exist in the default environment,
without overwriting protected variables.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-324