Revert "env: process -f (forcibly) option of 'env default' command"
After a discussion, it was decided that the side effect of this
patch (preserving variables which didn't exist in the default
environment) is undesired and can cause confusion.

I'm reverting this patch and try to solve it in a different way
so that the environment can be reset to its default state except
for the protected variables.

This reverts commit 72f63413066836d0589f9acb1c8c655c0ef6bbd1.