env: process -f (forcibly) option of 'env default <var>' command
The 'env default <var>' command offers the -f option to forcibly reset an
environment variable to its default value. This option is however not
taken into account with the following inconsistent behavior for protected
variables:

  env default ethaddr      => Does not change 'ethaddr' (OK)
  env default -f ethaddr   => Does not change 'ethaddr' (ERROR)

This patch propagates the -f flag (H_FORCE) to set_default_vars() to
solve this problem.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-324