env: process -f (forcibly) option of 'env default -a' command
The 'env default -a' command offers the -f option to forcibly reset the
whole environment to its defaults. This option is however not taken into
account with the following inconsistent behavior for protected variables:

  env default -a           => Changes/deletes 'ethaddr' (ERROR)
  env default -f -a        => Changes/deletes 'ethaddr' (OK)

This patch calls set_default_env() function with a star (*) as the first
character of the passed string if called with -f.
This is later processed to pass the flag H_FORCE to h_import() with
translates to reuse the hash table after removing all variables that don't
have a write-protect flag.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-324