ccimx6: use a unique variable 'video_args' to store the video cmdline
After discussion, we decided that having up to four videoN variables
for each framebuffer was overkilling and could cause confusion, as it
implicitly added the framebuffer string.
To let the user specify the video arguments completely, we've decided to
maintain only one variable $video_args which will be appended to the
command line for Android OS.

We have omitted it from Linux OS because Linux will define the video
settings directly on the Device Tree.

The default value of $video_args is to have HDMI on framebuffer 0.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-313