update: verify written data during OTF update mechanism
Changed the write_chunk() function to verify data written in chunk before
proceeding to next chunk. In the OTF mechanism we assume a chunk at
$loadaddr will never overwrite the read block at $verifyaddr.
The write_chunk() function does not return the number of read bytes but 0
if the write and verification was successful, or !=0 in case of error (be
it write/read/verify).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-288