update: fix OTF update when being called more than once
The OTF update hook started with correct data the first time it was called
but subsequent calls started with invalid offsets, which was causing
incorrect data to be written to the partition when two update operations
were run one after the other.
This commit introduces a new INIT flag, to signal the hook the need to
restart some of the internal variables. The flag is set during the hook
registration (on each 'update' operation) and reset after initializing
the variables.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-302