update: verify updated firmware by comparing 32-bit words instead of bytes
The standard update command was using the U-Boot command 'cmp.b' to
compare written data with read data. This doesn't benefit from the memory
bus width and slows down the update process.
Use 'cmp.l' instead to compare by 32bit words. To this end, we must first
pad with zeros potential bytes at the end, if the image is not a whole
multiple of 4.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>