ccimx6: rename 'std_bootargs' to 'extra_bootargs'
The contents of this variable will be appended to 'bootargs' before
booting with 'dboot' command.
In the past, the variable was called 'std_bootarg' without the 's'.
This name had no meaning and the lacking of an ending 's' lead to
confusion for users (standard variable is 'bootargs' with an 's').

To avoid the confusion and to make the meaning of the variable more
explicit, we have decided to rename it to 'extra_bootargs' which is
self explanatory.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-335