tftp: reduce the TFTP timeout to 1s
In all my experience, whenever a packet does not receive its
ACK, no matter how long you wait, you never receive it.
It does not make sense to wait such a long time as 5s.
Reduce it to 1s, to mimize the impact of timeouts in the overall
transfer time.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>