cmd_hwid: make manufstr_to_hwid() have a variable number of arguments
Manufacturing operations can now be given using several strings. Board
code will decide if the number and format are valid.
In the case of the CC6, apart from the 13 ASCII chars that carry the
manufacturing date and serial number, we also want to pass the module
variant, hardware version and wireless cert in the same operation to
complete the HWID information.
This is given as a second string VVHC where:
- VV is the variant, in hex
- H is the hardware version, in hex
- C is the wireles certification, in hex

Signed-off-by: Hector Palacios <hector.palacios@digi.com>