ccimx6: disable bootscript
Remove the 'bootscript' variable by default to disable the
automatic bootscript feature. The Micrel PHY may take more
than 5 seconds to complete autonegotiation when connected to
a switch, which causes a big delay in boot and the bootscript
feature to timeout.

A direct connection to a PC may reduce the autonegotiation to
a pair of seconds and make the feature feasible. This will be
documented.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-328
https://jira.digi.com/browse/DUB-330