ccimx6: force the PHY to master during negotiation depending on variable
Micrel PHY KSZ9031 takes up to 5 seconds to autonegotiate with Gigabit
switches. By setting the variable 'phy_mode' to "master" (default is
"auto") the PHY will be forced in master mode during master/slave
negotiation. This makes the autonegotiation faster but may fail when using
a direct cable connection or when the other end is a forced master too.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-328
https://jira.digi.com/browse/DUB-343