ccimx6: add function to read the carrier board version from OTP bits
Since the CC6 is an SMD module, it can be assembled on different carrier
boards. Carrier boards may suffer redesigns that require different
software behavior.
This patch adds a new function get_carrier_board_version() which reads out
the carrier board version coded in the OTP bits.
The OTP bits where the version is coded can be defined in the platform
header file using:
	CONFIG_CARRIERBOARD_VERSION_BANK
	CONFIG_CARRIERBOARD_VERSION_WORD
	CONFIG_CARRIERBOARD_VERSION_MASK
	CONFIG_CARRIERBOARD_VERSION_OFFSET

For the SBC carrier board, the version is stored at the lower 4 bits
of Bank4 Word6 (GP1).
A version value of 0 is considered an 'undefined version'.
The carrier board version is printed in the U-Boot banner together with
the platform name resulting, for example, in :

For version = 0
	Board: ConnectCore 6 SBC (undefined version)
For version = 1
	Board: ConnectCore 6 SBC v1

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-344