ccimx6sbc: power down the audio codec during startup
Starting at version 2 of the SBC, GPIO2_25 enables power to the SGTL5000
audio codec on the SBC board. This line is externally pull-up so by
default the audio codec is enabled. Since the codec does not have a reset
line, after a software reset it may be enabled and configured and spitting
noise. To avoid it, power down the codec by pulling the power enable line
down.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-325