tftp: abort OTF write operation if writing of a chuck failed.
If the OTF hook for writing a chunk fails, abort the TFTP operation.
Without this patch, the TFTP prints an error message but keeps on
transferring packets until the full file is transmitted. If a chunk fails
however it makes no sense to keep on transmitting the file and the
operation must be aborted.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-334