hwid: add command to read/write the carrier board version from OTP
Although not part of the HWID itself, the carrier board version has many
similarities to the HWID and could reuse some functions from the
cmd_hwid.c file.

The patch adds a new 'board_version' command wit four subcommands:
 - read
 - sense
 - prog
 - override
which behave similar to the 'fuse' and 'hwid' commands.
The prog/override subcommands accept just one parameter: the carrier
board version which must be given in decimal.
The reason to use a decimal number (versus default hexadecimal in U-Boot)
is to match the carrier board version number (which will be silkprinted
in the board). In any case, it will be odd to have more than nine versions
of a carrier board.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-345