ARM: mx6: allow to boot when CPU temp is less or equal to a configured value
U-Boot had a hardcoded figure of 80ºC above which it refuses to boot, to
let the CPU cool down.
Make this temperature limit configurable as a value below the max CPU die
temperature coded in the OTP bits (which varies depending on the CPU
model).

For the cc6 boards, set this limit as 10ºC below the max CPU's die temp.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-347