ccimx6: early init function for ccimx6 to read the HWID
This patch adds function ccimx6_early_init() that is called from
board_early_init_f() function. This function is meant to hold early
initialization of the cc6 module, including reading of the HWID which
we may need to do early to be able to have conditional code basing on
the HWID values.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>