ccimx6: dynamically generate variable $verifyaddr
Variable $verifyaddr is used to read back a chunk of data during the
update command (both in normal and OTF modes) to do the verification
of the written data.
The value was originally hardcoded half way through the RAM starting
at $loadaddr, which was fine for 1GiB variant but wrong for variants
with less RAM.

This patch removes the variable and creates it dynamically if it doesn't
previously exist.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-391
https://jira.digi.com/browse/DUB-392