ccimx6: use constants for DDR configuration and split it in files
DDR IOMUX and configuration was done in just one file using address
values. This made it difficult to reuse, especially when building for
a different i.MX6 family like the DualLite/Solo.
This commit changes the address values with constants from headers
whose value differs basing on the selected CPU family.
It also splits the configuration in different files:
- ddr-setup.cfg: for common DDR iomux setup
- ccimx6_4x2Gbit.cfg: for the DDR configuration especific to a DDR layout
- clocks.cfg: for common clocks configuration

The platform refers to another file called after the variant:
- ccimx6q1GB.cfg: this file basically includes the other three

Signed-off-by: Hector Palacios <hector.palacios@digi.com>