update: auto activate on-the-fly mechanism when there is not enough RAM
There is no way to know in advance the size of the firmware file that is
going to be transferred, so an update command could be writing to RAM
beyond the available size, overriding U-Boot and hanging the target.
We know, however, the max size of the target partition the firmware
is going to be written to. If the available RAM (starting at $loadaddr
and ending before U-Boot location at the end of the RAM) is smaller than
the size of the target partition, activate on-the-fly mechanism
automatically (unless $otf-update is set to "no").

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-362
(cherry picked from commit 548497ee7f399e628bf46bed7861b65a9ad9aec8)