update: calculate $verifyaddr during 'update' command
If the variable $verifyaddr is not set, its value was calculated as half
way between the $loadaddr and the end of the RAM.
This was done in board late init code, instead of being done in a common
place. Besides, if the variable was deleted after booting, the verify
process did not work.

This patch moves the calculation of $verifyaddr to the 'update' command
thus accounting for potential deletions of the variable.
The code has also been fixed to use $loadaddr in the calculation instead
of hardcoded CONFIG_LOADADDR.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-425
(cherry picked from commit 52398d30055a505b713a7c369a8f406a54b02f3b)