dboot: correct unaccuracy in command's help
dboot is able to boot from alternative partitions, but not as it's said
in the command's help. The first parameter needs to be the reserved word
for the operating system and not anything related with the partition.
This is then used to compose the correct boot arguments to the kernel.

To boot from alternative partitions you just need to set the
rootfs-to-boot in the 'mmcroot' variable and pass the boot partition in
the 'dboot' command:

For example, to boot from the second FAT partition and mount the second
rootfs partition 'rootfs2' as main rootfs:

setenv mmcroot "PARTUUID=c7d8648b-76f7-4e2b-b829-e95a83cc7b32"
dboot linux mmc 0:2

So just remove the misleading part of the command's help.

https://jira.digi.com/browse/DUB-389

Signed-off-by: Javier Viguera <javier.viguera@digi.com>