ccimx6: move arrays definitions from header to board code
In order to reuse the variants array in the carrier board code, we need
to define it in the SOM code and declare it as extern in the header file.
At the same time, we move the definitions of capabilities below the
variants struct, and move the cert_regions definition also to the SOM code.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit 134b4cc368e4199c91e0fc548f7bdbb493f8bee2)