cmd_update: remove hard coded index for storage media
Up to now, the update command worked with constant
CONFIG_SYS_STORAGE_DEV to determine the index of the storage
device. There are now CC6 variants that don't have eMMC and
whose storage device will be the uSD card, but we also want to
be able to use update over a removable media like uSD card from
a variant with eMMC.

This commit creates a function mmc_get_bootdevindex() that
returns the index of the boot device. If $mmcdev is undefined,
it will be set to the boot media. Then the update command will
write to the device pointed to by $mmcdev.

Global variables 'mmc_dev' and 'mmc_dev_index' are used to point to the
system storage device through OTF-update board hooks.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-320
(cherry picked from commit a88825fe0994742abad37d8456401e7481f481e5)