ccimx6: if invalid HWID, assume full populated variant
If an invalid variant is coded on the HWID (or nothing is coded
at all) the USDHC4 controller will not be initialized, preventing
access to the eMMC from U-Boot.
This can cause trouble to manufacturing/test or to customers
that accidentally receive a hardware without HWID.
This commit assumes a fully populated variant if an invalid HWID
is read on the OTP bits, which will allow the hardware to work
in this scenario, if it is really there.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-442
(cherry picked from commit db7f599906dd7dcf2ad057b916bdabb301664a07)

Conflicts:
  board/digi/ccimx6/ccimx6.c