ccimx6: set default values for $mmcdev and $mmcbootpart and not overwrite them
Overwriting these variables by U-Boot was considered a bad idea and we
decided to set them to default values of 0 and CONFIG_SYS_BOOT_PART_EMMC
(1), which covers most of the use cases.
For the use case of modules with eMMC wanting to update to the uSD card the
user needs to manually set mmcdev to 1.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-445
(cherry picked from commit 8ba0183c866ca540b7a61248a2647ef89d5a17d2)